requirements.txt
aspose-words==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
docx2pdf==0.1.8
docxtpl==0.11.4
et-xmlfile==1.1.0
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
numpy==1.23.4
openpyxl==3.0.9
pandas==1.3.5
Pillow==9.2.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2022.5
pywin32==304
requests==2.28.1
six==1.16.0
tqdm==4.64.1
urllib3==1.26.12
Werkzeug==2.2.2
